+++ CHANGES.md
@@ -67,6 +67,7 @@ To be released.
  -  Removed `HashCash` class.  [[#3067]]
  -  (@planetarium/account) Replaced `Account.publicKey` property with
     `Account.getPublicKey()` async method.  [[#3061]]
+ -  (@planetarium/account) Added `Account.getAddress()` async method.  [[#3084]]
  -  (@planetarium/account) `Address.deriveFrom()` method now returns
     `Promise<Address>` when an `Account` is given.  However, it still returns
     `Address` when a `PublicKey` is given.  [[#3061]]
@@ -110,12 +111,17 @@ To be released.
  -  (Libplanet.Net) Added `Gossip.PublishMessage()` method.  [[#3054], [#3060]]
  -  (@planetarium/account) Added `Account.getPublicKey()` async method.
     [[#3061]]
+ -  (@planetarium/account) Added `Account.getAddress()` async method.  [[#3084]]
  -  (@planetarium/account) Added `RawPrivateKey.getPublicKey()` async method.
     [[#3061]]
+ -  (@planetarium/account) Added `RawPrivateKey.getAddress()` async method.
+    [[#3084]]
  -  (@planetarium/account-aws-kms) Added `AwsKmsAccount.getPublicKey()` async
     method.  [[#3061]]
+ -  (@planetarium/account-aws-kms) Added `AwsKmsAccount.getAddress()` async
+    method.  [[#3084]]
  -  (@planetarium/account-web3-secret-storage) Added `Web3Account` class.
-    [[#3061]]
+    [[#3061], [#3084]]
  -  (@planetarium/account-web3-secret-storage) Added `Web3KeyObject` interface.
     [[#3061]]
  -  (Libplanet.Explorer) Added several interfaces and classes that pertain to
@@ -255,6 +261,7 @@ To be released.
 [#3077]: https://github.com/planetarium/libplanet/pull/3077
 [#3079]: https://github.com/planetarium/libplanet/pull/3079
 [#3080]: https://github.com/planetarium/libplanet/pull/3080
+[#3084]: https://github.com/planetarium/libplanet/pull/3084
 
 
 Version 0.53.4